MGS-4724 Use new govscot:slug property to control publication urls
In order to ensure that each publication has a unique url we have
added the govscot:slug property to all publications types:
  * govscot:Publication
  * govscot:Minutes
  * govscot:SpeechOrStatement
  * govscot:ComplexDocument
This property is required and has custom validation to ensure that it
is unique (PublicationSlugValidator).

When a new publication is created we automatically assign a new slug to
it based on its folder name disambiguated by adding a number if needed.
(PublicationSlugDaemonModule).

The link processor for publications has also been updated to use this
property both when generating links and in interpreting urls.