MGS-5760 404 page now indicates if the content has been archived
In order to allow us to track what archived content is still being used
we now serve a 404 page with a message to indicate that the content has
been archived.

The 'archive' url is now content editabel from this page.  This will
allow us to switch from www2 to the archive when the pulsant contract
ends.

This also requries that the migration New404Page be run - this changes
the type of the 404 page to that wee can make the archive title, url
and content editable by cms-admins.